fix(secretsmanager): resolve secrets by partial ARN prefix match — bu… (#268)
* fix(secretsmanager): resolve secrets by partial ARN prefix match — bump v1.2.5

* fix cross compat SDK

* adding fix for empty body on GET server

+++ ministack/services/kms.py
@@ -187,7 +187,7 @@ def _create_key(data):
         "KeySpec": key_spec,
         "KeyUsage": key_usage,
         "Description": description,
-        "CreationDate": time.time(),
+        "CreationDate": int(time.time()),
         "Origin": "AWS_KMS",
         "Tags": tags,
         "Policy": policy,
@@ -764,7 +764,7 @@ def _schedule_key_deletion(data):
     days = data.get("PendingWindowInDays", 30)
     rec["KeyState"] = "PendingDeletion"
     rec["Enabled"] = False
-    rec["DeletionDate"] = time.time() + (days * 86400)
+    rec["DeletionDate"] = int(time.time() + (days * 86400))
     return json_response({
         "KeyId": rec["KeyId"],
         "KeyState": "PendingDeletion",
